+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version: 0.27.33
+
+### New
+- Added ability to call `account` command with address from config
+
+### Big fixes
+- Fixed work with old config file
+
+
+## Version: 0.27.31
+
+### New
+ - Clear alternative syntax parameters
+ - Alias and abi methods completion
+
+
 ## Version: 0.27.30
 
 ### New